feat: [FE] 상품 리스트에서 상품 이미지 노출
