+++ delivery-market-admin/src/main/java/com/ddangme/dmadmin/controller/goods/GoodsApiController.java
@@ -2,18 +2,17 @@ package com.ddangme.dmadmin.controller.goods;
 
 import com.ddangme.dmadmin.dto.Response;
 import com.ddangme.dmadmin.dto.goods.request.GoodsSaveRequest;
-import com.ddangme.dmadmin.model.constants.UploadFile;
 import com.ddangme.dmadmin.service.FileUploadService;
 import com.ddangme.dmadmin.service.goods.GoodsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.net.MalformedURLException;
 
 @RestController
 @Slf4j
@@ -22,6 +21,7 @@ import java.io.IOException;
 public class GoodsApiController {
 
     private final GoodsService goodsService;
+    private final FileUploadService fileUploadService;
 
     @PostMapping("/add")
     public Response<Void> add(GoodsSaveRequest request, @RequestParam(required=false) MultipartFile photo) throws IOException {
@@ -30,5 +30,10 @@ public class GoodsApiController {
 //        goodsService.save(request.toDTO(), photo);
 
         return Response.success();
+    }
+
+    @GetMapping("/images/{filename}")
+    public Resource downloadImage(@PathVariable String filename) throws MalformedURLException {
+        return new UrlResource("file:" + fileUploadService.getFullPath(filename));
     }
 }
