+++ delivery-market-admin/src/main/resources/templates/goods/goods-add.html
@@ -202,7 +202,7 @@
                 </div>
             <div class="d-flex justify-content-center mt-5">
                 <button type="submit" class="btn btn-outline-success" id="add-category-btn">추가</button>
-                <a type="button" class="btn btn-outline-secondary ms-3" href="/categories">취소</a>
+                <a type="button" class="btn btn-outline-secondary ms-3" href="/goods">취소</a>
             </div>
             </form>
 
@@ -220,4 +220,4 @@
 
 <script>
 </script>
-<script src="/js/goods.js"></script>
+<script src="/js/goods-add.js"></script>
