+++ delivery-market-admin/src/main/resources/templates/goods/goods-list.html
@@ -39,7 +39,7 @@
         </div>
         <div class="col-12">
             <div class="table-responsive small">
-                <table class="table table-striped table-sm text-center">
+                <table class="table table-hover table-sm text-center">
                     <thead>
                     <tr>
                         <th scope="col">
@@ -50,33 +50,35 @@
                         <th scope="col">상품명</th>
                         <th scope="col">가격</th>
                         <th scope="col">판매 상태</th>
-                        <th scope="col">업로드 이미지명</th>
-                        <th scope="col">서버 저장 이미지명</th>
+                        <th scope="col">대표 이미지</th>
                         <th scope="col">생성일자</th>
                         <th scope="col">생성자</th>
                         <th scope="col">수정일자</th>
                         <th scope="col">수정자</th>
                     </tr>
                     </thead>
-                    <tbody th:each="good: ${goods}">
-                        <tr>
-                            <td>
-                                <input class="form-check-input" type="checkbox">
-                            </td>
-                            <td th:text="${good.id}"></td>
-                            <td th:text="${good.categoryName}"></td>
-                            <td th:text="${good.name}"></td>
-                            <td th:text="${good.price}"></td>
-                            <td th:text="${good.saleStatus.getStatus()}"></td>
-                            <td th:text="${good.uploadFileName}"></td>
-                            <td th:text="${good.storeFileName}"></td>
-                            <td th:text="${#temporals.format(good.createdAt, 'yyyy-MM-dd HH:mm:ss')}"></td>
-                            <td th:text="${good.createdBy}"></td>
-                            <td th:if="${good.createdAt == good.updatedAt}"></td>
-                            <td th:if="${good.createdAt != good.updatedAt}" th:text="${#temporals.format(good.updatedAt, 'yyyy-MM-dd HH:mm:ss')}"></td>
-                            <td th:if="${good.createdAt == good.updatedAt}"></td>
-                            <td th:if="${good.createdAt != good.updatedAt}" th:text="${good.updatedBy}"></td>
-                        </tr>
+                    <tbody>
+                        <div th:each="good: ${goods}">
+                            <tr class="clickable-row" th:id="${good.id}">
+                                <td>
+                                    <input class="form-check-input" type="checkbox">
+                                </td>
+                                <td th:text="${good.id}"></td>
+                                <td th:text="${good.categoryName}"></td>
+                                <td th:text="${good.name}"></td>
+                                <td th:text="${good.price}"></td>
+                                <td th:text="${good.saleStatus.getStatus()}"></td>
+                                <td>
+                                    <img th:src="|/api/goods/images/${good.getStoreFileName()}|" width="auto" height="30">
+                                </td>
+                                <td th:text="${#temporals.format(good.createdAt, 'yyyy-MM-dd HH:mm:ss')}"></td>
+                                <td th:text="${good.createdBy}"></td>
+                                <td th:if="${good.createdAt == good.updatedAt}"></td>
+                                <td th:if="${good.createdAt != good.updatedAt}" th:text="${#temporals.format(good.updatedAt, 'yyyy-MM-dd HH:mm:ss')}"></td>
+                                <td th:if="${good.createdAt == good.updatedAt}"></td>
+                                <td th:if="${good.createdAt != good.updatedAt}" th:text="${good.updatedBy}"></td>
+                            </tr>
+                        </div>
                     </tbody>
                 </table>
             </div>
